aider: Add check to ensure `folder_path` is defined before opening it.
# Aider chat conversation:

USER: Clicking The process Images but should als copy the imaeges I also get this error Good morning,

Thank you for your interest in the property in Old Burrin. Carlow.

We received a huge response for this property. Unfortunately, on this occasion we cannot accommodate a viewing for you as we are limited on the amount of viewings we can arrange for this property.

Apologies for any inconvenience.
Good luck with your search, please keep an eye on daft.ie and if you see anything you are interested in please get in touch with us.

Kind regards,
Maher Property Advisors
Licence No: 003837
0599143100

• Sales • Lettings • Commercial
www.maherproperty.ie
The contents of this e-mail are Subject to Contract /Contract Denied. This email and any files transmitted with it are confidential and intendedsolely for the use of the individual or entity to whom they are addressed. If you have received this email in error please notify us immediately at bernadette@maherproperty.ie and delete this email from your system. Thank you. It is possible for data transmitted by email to be deliberately or accidentally corrupted or intercepted.  For this reason, where communication is by email, Maher Property Advisors does not accept responsibility for any breach of confidence which may arise through the use of this medium.
Subject to contract /contract denied

ASSISTANT: It seems like you're experiencing an issue with the "Process Images" button. However, the error message you provided doesn't seem to be related to the software. It looks like an email response from a property advisor.

Could you please provide more details about the error you're experiencing? Specifically, what happens when you click the "Process Images" button? What error message do you see? This will help me understand the issue better and provide a more accurate solution.

USER: Exception has occurred: NameError
name 'folder_path' is not defined
  File "C:\Users\SoftCroft\Documents\GitHub\photos_maher\rename_camera2.py", line 126, in open_location_callback
    os.startfile(folder_path)
                 ^^^^^^^^^^^
  File "C:\Users\SoftCroft\Documents\GitHub\photos_maher\rename_camera2.py", line 122, in _process_images_thread
    self.open_location_callback(True)  # Open the location
    ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
NameError: name 'folder_path' is not defined

ASSISTANT: The error message indicates that the variable `folder_path` is not defined at the time when the `open_location_callback` method is called. This could happen if the "Process Images" button is clicked before a folder is created using the "Create Folder" button.

To fix this, we can add a check in the `open_location_callback` method to ensure that `folder_path` is defined before trying to open it. If `folder_path` is not defined, we can show an error message to the user.

Here is the change:

```python
rename_camera2.py
<<<<<<< HEAD
    def open_location_callback(self, response):
        if response:
            os.startfile(folder_path)
=======
    def open_location_callback(self, response):
        global folder_path
        if response:
            if 'folder_path' in globals():
                os.startfile(folder_path)
            else:
                messagebox.showerror("Error", "No folder has been created yet.")
>>>>>>> updated
```

This change adds a check to see if `folder_path` is defined. If it is, the folder is opened. If it's not, an error message is shown to the user.