aider: Initialized `folder_path` as a global variable and updated it in the `create_folder` method.
# Aider chat conversation:

USER: when I click process image it says no folder has been created yet ecevn though it has

ASSISTANT: The issue seems to be related to the scope of the `folder_path` variable. It's defined as a global variable in the `create_folder` method, but it's not recognized as such in the `open_location_callback` method. This could be due to the fact that the `create_folder` method might not have been called before the `open_location_callback` method, which would mean that the `folder_path` variable hasn't been defined yet.

To fix this, we can initialize `folder_path` as a global variable at the start of the script, and then update it in the `create_folder` method. This way, `folder_path` will always be defined, even if the `create_folder` method hasn't been called.

Here is the change:

```python
rename_camera2.py
<<<<<<< HEAD
class App(ctk.CTk):
    def __init__(self):
        super().__init__()
=======
folder_path = None

class App(ctk.CTk):
    def __init__(self):
        super().__init__()
        global folder_path
>>>>>>> updated
```

```python
rename_camera2.py
<<<<<<< HEAD
    def create_folder(self):
        global folder_path
        property_name = self.property_name_entry.get()
        folder_path = os.path.join(base_folder_path, property_name)
        os.makedirs(folder_path, exist_ok=True)
        response = messagebox.askyesno("Folder created", "Folder created. Do you want to process images?")
        self.create_folder_callback(response)
=======
    def create_folder(self):
        global base_folder_path, folder_path
        property_name = self.property_name_entry.get()
        folder_path = os.path.join(base_folder_path, property_name)
        os.makedirs(folder_path, exist_ok=True)
        response = messagebox.askyesno("Folder created", "Folder created. Do you want to process images?")
        self.create_folder_callback(response)
>>>>>>> updated
```

This change initializes `folder_path` as a global variable at the start of the script, and then updates it in the `create_folder` method.