aider: Removed the `sticky="ew"` option from the `grid` method for `self.logo_label`.
# Aider chat conversation:

USER: I can see the text `CTkLabel` in the image and the image is barely visable now

ASSISTANT: It seems like the `CTkLabel` is being displayed on top of the image. This could be due to the way the `CTkLabel` is being packed or gridded in the tkinter layout.

To fix this, we can try to adjust the layout of the `CTkLabel` and the image.

Here's the edit block to make this change:

```python
rename_camera2.py
<<<<<<< HEAD
self.logo_label = ctk.CTkLabel(self, image=self.logo_image)
self.logo_label.grid(row=4, column=0, columnspan=4, padx=20, pady=10, sticky="ew")
=======
self.logo_label = ctk.CTkLabel(self, image=self.logo_image)
self.logo_label.grid(row=4, column=0, columnspan=4, padx=20, pady=10)
>>>>>>> updated
```

This change removes the `sticky="ew"` option from the `grid` method for `self.logo_label`. The `sticky` option is used to specify which sides of the cell the widget should stick to. The value `"ew"` means the widget will stick to the east and west (left and right) sides of the cell, which could be causing the `CTkLabel` to cover the image. By removing this option, the `CTkLabel` should only take up as much space as it needs, allowing the image to be displayed properly.