aider: Resize the logo image to 600x300 pixels in the `rename_camera2.py` file.
# Aider chat conversation:

USER: CTkImage
The CTkImage is not a widget itself, but a container for up to two PIL Image objects for light and dark mode. There's also a size tuple which describes the width and height of the image independent of scaling. Therefore it's important that the PIL Image's are in a higher resolution than the given size tuple, so that the image is not blurry if rendered on a 4K monitor with 2x scaling. So that the image is displayed in sharp resolution on a 2x scaled monitor, the given OIL Image's must have at least double the resolution than the requested size.

Create a CTkImage object:

from PIL import Image

my_image = customtkinter.CTkImage(light_image=Image.open("<path to light mode image>"),
                                  dark_image=Image.open("<path to dark mode image>"),
                                  size=(30, 30))

image_label = customtkinter.CTkLabel(app, image=my_image, text="")  # display image with a CTkLabel

Arguments
argumentvalue
light_imagePIL Image object for light mode
dark_imagePIL Image object for dark mode
sizetuple (width in px, height in px) for rendering size independent of scaling
If only light_image or only dark_image is given, the existing one will be used for both light and dark mode.

Methods
.configure(attribute=value, ...)
All attributes can be configured and updated.

.cget(attribute_name)
Pass attribute name as string and get current value of attribute.

ASSISTANT: I understand that you're providing information about the `CTkImage` class from the `customtkinter` library. However, I'm not sure what you're asking for. Are you experiencing an issue with the `CTkImage` class? Or do you want to make changes to your code related to this class? Could you please provide more details or clarify your request?

USER: I can see the log image at the bottom can you use that

ASSISTANT: I see that you want to use the `CTkImage` class to display the logo image at the bottom of the application. However, the image is already being displayed using the `CTkImage` class in your current code.

If you're asking to change the size of the image, we can do that by modifying the size parameter when resizing the image.

For example, if you want to change the size of the image to 600x300 pixels, we can modify the `img.resize` line in the `__init__` method of the `App` class in the `rename_camera2.py` file.

Here is the edit block for that change:

```python
rename_camera2.py
<<<<<<< HEAD
        img = img.resize((400, 200), Image.LANCZOS)  # Resize the image to 500x500 pixels
=======
        img = img.resize((600, 300), Image.LANCZOS)  # Resize the image to 600x300 pixels
>>>>>>> updated
```

Please confirm if this is the change you want to make.