aider: Changed image resizing method to Image.BICUBIC for better quality.
# Aider chat conversation:

USER: remeber this didnt work as image.Antialias is not availble

ASSISTANT: I apologize for the confusion. The `Image.ANTIALIAS` filter is not available in your version of the PIL library. Let's try a different approach. We can use the `Image.BICUBIC` filter which is a good filter for resizing images.

Here is the *edit block* for the change:

```python
rename_camera2.py
<<<<<<< HEAD
        img = Image.open("maher.png")
        img = img.resize((600, 300), Image.ANTIALIAS)  # Resize the image to 600x300 pixels
        self.logo_image = CTkImage(img)
=======
        img = Image.open("maher.png")
        img = img.resize((600, 300), Image.BICUBIC)  # Resize the image to 600x300 pixels
        self.logo_image = CTkImage(img)
>>>>>>> updated
```

In the updated code, I've changed the resizing method to `Image.BICUBIC` which should provide a good quality resizing for your image.